from 1080

+++ src/caffe/util/im2col_cl.cpp
@@ -60,11 +60,11 @@ template void im2col_gpu<float>(const float* data_im, const int channels,
     const int pad_h, const int pad_w, const int stride_h,
     const int stride_w, const int dilation_h, const int dilation_w,
     float* data_col);
-template void im2col_gpu<double>(const double* data_im, const int channels,
+template void im2col_gpu<half>(const half* data_im, const int channels,
     const int height, const int width, const int kernel_h, const int kernel_w,
     const int pad_h, const int pad_w, const int stride_h,
     const int stride_w, const int dilation_h, const int dilation_w,
-    double* data_col);
+    half* data_col);
 
 template <typename Dtype>
 inline void im2col_nd_core_gpu(const Dtype* data_input, const bool im2col,
@@ -152,10 +152,10 @@ template void im2col_nd_gpu<float>(const float* data_im, const int num_spatial_a
     const int col_size, const int* im_shape, const int* col_shape,
     const int* kernel_shape, const int* pad, const int* stride,
     const int* dilation, float* data_col);
-template void im2col_nd_gpu<double>(const double* data_im, const int num_spatial_axes,
+template void im2col_nd_gpu<half>(const half* data_im, const int num_spatial_axes,
     const int col_size, const int* im_shape, const int* col_shape,
     const int* kernel_shape, const int* pad, const int* stride,
-    const int* dilation, double* data_col);
+    const int* dilation, half* data_col);
 
 template <typename Dtype>
 void col2im_gpu(const Dtype* data_col, const int channels,
@@ -200,11 +200,11 @@ template void col2im_gpu<float>(const float* data_col, const int channels,
     const int pad_h, const int pad_w, const int stride_h,
     const int stride_w, const int dilation_h, const int dilation_w,
     float* data_im);
-template void col2im_gpu<double>(const double* data_col, const int channels,
+template void col2im_gpu<half>(const half* data_col, const int channels,
     const int height, const int width, const int kernel_h, const int kernel_w,
     const int pad_h, const int pad_w, const int stride_h,
     const int stride_w, const int dilation_h, const int dilation_w,
-    double* data_im);
+    half* data_im);
 
 template <typename Dtype>
 void col2im_nd_gpu(const Dtype* data_col, const int num_spatial_axes,
@@ -221,10 +221,10 @@ template void col2im_nd_gpu<float>(const float* data_col, const int num_spatial_
     const int im_size, const int* im_shape, const int* col_shape,
     const int* kernel_shape, const int* pad, const int* stride,
     const int* dilation, float* data_im);
-template void col2im_nd_gpu<double>(const double* data_col, const int num_spatial_axes,
+template void col2im_nd_gpu<half>(const half* data_col, const int num_spatial_axes,
     const int im_size, const int* im_shape, const int* col_shape,
     const int* kernel_shape, const int* pad, const int* stride,
-    const int* dilation, double* data_im);
+    const int* dilation, half* data_im);
 
 
 }  // namespace caffe
